+++ NEWS.md
@@ -1,3 +1,13 @@
+# tidychangepoint 1.0.3
+
+* Added support for models from the `strucchange` package.
+* Added group aesthetic to plots (@rpruim)
+* The dots (...) are passed more consistently
+* Moved `patchwork` to Suggests
+* Moved `stats` reexports to imports
+* Moved `vctrs` reexports to imports
+* Moved `broom` reexports to `generics`
+
 # tidychangepoint 1.0.2
 
 * Updated use of `changepointGA::GA()` to `changepointGA::cptga()`